refactor: Simplify VirtioDevice trait by combining methods
Combine the `interrupt_evt` and `interrupt_status` methods into a single
method `interrupt_trigger` that returns a `IrqTrigger` reference (which
essentially combines the two objects originally returned by the status
and evt methods).

The advantage to this is that `IrqTrigger` exposes a `trigger_irq`
method, which I'd like to use in the next commit.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>