doc: Update guest kernel patch version in getting-started.md
Fix the whole "we forgot to update the getting started guide after
regenerating CI artifacts" thing once and for all by using S3's
list-objects API for dynamically looking up the name of the 5.10 guest
kernel. We commit the cardinal sin of parsing XML using regex, but
this should be fine here as we're only matching for a single, complete
tag. This breaks if the list objects operation ever returns more than
1000 objects, but that would meant over 1000 5.10 guest kernels in our
CI artiafacts, and I doubt that will ever happen.

Closes #4675

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>