virtio-net: correctly handle IPv6 TSO when supported
In the past, we were enabling IPv6 TSO in the TAP device but we were not
negotiating the flag with the guest driver. The previous commit cleaned
this up, by not setting gen::TUN_F_TSO6 when calling the TUNSETOFFLOAD
ioctl. This commit properly negotiates the feature and, if the guest
supports it, it enables it in the TAP device.

Signed-off-by: Babis Chalios <bchalios@amazon.es>