fix: use `version = 2` in `deny.toml`
New versions of `cargo deny` default all warnings to deny by default,
and removed the possibility to configure them (e.g. the `unlicensed` and
`copyleft` options are removed, with the default behavior now being as
if they had been set to `deny` before).

We opt-in to this new behavior by setting `version = 2` and removing the
now unsupported configuration options, to avoid a warning from being
printed (which would cause our `test_licences` test to fail).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>