fix: disable `clippy::cast_lossless` lint
With the new toolchain, this lint fires inside macros, such as
`arm64_core_reg_id`. However, this macro is uses in `const` contexts,
and the suggested replacement `u64::from` is not a const function, so is
not applicable. Furthermore, we also cannot suppress the lint here, as
annotating expressions (such as the body of the `arm64_core_reg_id`
macro) with attributes is not allowed on stable compilers. Thus, it is
impossible to address this warning.

We can safely disable this specific lint though, as it is not gaining us
anything anyway: It explicitly disallows lossless conversions, e.g.
those that are always safe. The justification here is that "if types
ever change, these conversions can become silently lossly", however that
does not apply in our case, as we explicitly disable all lossly
conversions. Thus, disable this lint.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>