Edit CPUID Normalization to account for possibility of hotplugging
Changes some of the CPUID leaves to account for the possibility of
hotplugged vCPUs. Previously, the maximum APIC address space and L3
cache were only shared between boot_cpus. Whereas, now we want this to
be extended for the maximum number of CPUs that can be added,
MAX_SUPPORTED_VCPUS.

Signed-off-by: James Curtis <jxcurtis@amazon.co.uk>