ci: patch ARM 6.1 guest kernel with VMGenID support
Linux support for VMGenID on ARM systems landed in kernel 6.10. We only
support up to kernel 6.1. So in order to be able to test VMGenID on ARM
systems we need to patch our 6.1 kernel to backport the 6.10 commits.

This commit adds the patches in our repo under
`resources/patches/vmgenid_dt` and changes rebuild.sh to patch 6.1
kernel on ARM before building it. Also, it enables `CONFIG_VMGENID` for
ARM 6.1 guest kernel.

Eventually we will switch to using guest kernels from Amazon Linux trees
hosted on Github, which they will eventually backport the relevant
patches. Once this is done we can drop the patches from our tree and we
won't need to manually patch them.

Signed-off-by: Babis Chalios <bchalios@amazon.es>