vmgenid: create new DTB files for unit tests
Our unit tests for creating FDT depends on DTB blobs that we create once
and then we use as a reference to ensure that we don't make a change
that breaks them.

Now that we pass VMGenID via DeviceTree we need to recreate these blobs,
because we modify the layout of the guest memory, so the memory node is
suddenly different.

Also, fix the code that we use to create these blobs. Since, we
re-organized crates in the repo, it was trying to write to non-existent
paths.

Signed-off-by: Babis Chalios <bchalios@amazon.es>