feat: add support for VIRTIO_NET_F_MRG_RXBUF to virtio-net
Now virtio-net device can split incoming packets across
multiple descriptor chains if VIRTIO_NET_F_MRG_RXBUF is enabled
by the guest. The amount of descriptor chains (also known as heads)
is written into the `virtio_net_hdr_v1` structure which is located
at the very begging of the packet. Virtio spec states that the
number of heads used should always be correct:
- 1 - if VIRTIO_NET_F_MRG_RXBUF is not negotiated
- N - if VIRTIO_NET_F_MRG_RXBUF is negotiated
Prior to this commit Firecracker never set the number of
used heads to 1, but Linux was fine with it. Now we always
set correct number of heads. Because of this some changes
were introduced into the unit test code that was generating
testing frames.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>