vmgenid: allocate only 16 bytes for VMGenID
Following the QEMU implementation, we were allocating a full page for
holding the memory of the VMGenID device, even though the device itself
is only using 16 bytes for the generation ID.

We actually don't have any reason to do that. The Linux kernel itself is
just remapping 16 bytes from the address we pass over via
ACPI/DeviceTree.

Revert to allocating 16 bytes for the VMGenID device and expose the
value via a public constant for other systems to consume.

Signed-off-by: Babis Chalios <bchalios@amazon.es>