fix(test): do not use build step for optional pipeline
Optional pipeline started failing during a build failure in the rust
benchmarks tests.

During the build step, we build REVISION_A (main) and REVISION_B (tip of
the PR commit) and pass it over to the subsequent pipeline steps that
run the test.

However, 'test_benchmarks.py' needs to build the benchmarks, running
`cargo bench`. That step started failing when building 'aws-lc-rs-sys'
because CMake is complaining that its CMakeCache.txt file is not in the
same path as the one initially created.

Since we anyway run `cargo bench` in the test there's not really need to
pre-build firecracker. So, modify the buildkite pipeline logic to allow
defining pipelines without the build step, and change the
pipeline_pr_no_block.py to not use it.

Signed-off-by: Babis Chalios <bchalios@amazon.es>