test: cause failure if FC emits an error metric
When flushing firecracker metrics to cloudwatch, explicitly assert that
metrics containing the keypharses "err", "fail" or "panic" are all 0.

While doing this, clean up the flushing code a bit by recognizing that
if we separate the flattening of the metrics json and the actual
emitting, we can treat the flattening as a tree-walk (since we are
converting a tree to a set of (path, leaf) pairs).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>