test: Move `FcMetricsMonitor` initialization into `Microvm`
This way, the lifecycle of the monitor perfectly matches that of the
firecracker process (as it is started up when the Firecracker process
starts, and gets terminated when the microvm is killed).

This also means that the assertion the previous commit introduces will
only trigger during teardown (or when explicitly killing microVMs), at
which will be _after_ all other test-specific metrics have been emitted
(which means that triggering the assertion will not lead to loosing
performance metrics).

Also just start the monitor thread in the constructor, since all
call-sites just immediately started it anyway.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>