test(rpc_interface): Eliminate mocking
The remaining tests all pass without changes if we swap out the mock
framework for an actual Vmm/VmResources pair. While this is probably not
a good fit for "unittesting", we already do this "spin up an actual VMM
inside a unittest" all over the place, and fixing that would be a
significantly larger endeavor.

By eliminating the mocking, we get rid of another instance of
`#[cfg(not(test))]/#[cfg(test)]` in production code. This one is
particularly nice to get rid of because it secretly replaced the
definition of the Vmm/VmResources structs when tests were run, much to
the confusion of my IDE.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>