refactor(test): Don't use closures for checking return values
The tests in rpc_interface.rs were using a pattern where the result of
an API request was passed to a closure, along with a reference to the
VMM object (or a similar related object). However, by noticing that the
reference is not used in _any_ of the tests, we can drop this parameter
from the closure. But then we can also go a step further and just return
the `Result<VmmData, VmmActionError>` from the utility methods, to have
a more idiomatic assertion pattern.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>