test(jailer): do not kill screen with new_pid_ns
In our setup, if Firecracker is run by the screen
with --new-pid-ns passed to the jailer,
the Firecracker process detaches from the screen
causing the screen process to exit as well.
We attempt to kill the screen process by pid,
which causes a race condition.

This change lets the screen terminate on its own
without attempting to kill it.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>