fix: avoid potential panic when restoring malformed MMDS
If a snapshot file is malformed in a way that a field that is expected
to never be zero ends up being zero, Firecracker would panic. Fix this
by consistency typing this field as `NonZeroUsize` through the code
base. For the `NonZero*` types, serde already has support. Now if a
snapshot file is malformed and has zeroes here, serde will return an
error, which causes Firecracker to exit gracefully instead of panicking.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>