fix: do not panic if malformed snapshot file contains vhost devices
If a snapshot file is corrupted in a way that makes it _look_ like it
contains a vhost-user-blk device, restoring will hit an `unimplemented!`
inside `block::vhost_user/persist.rs`. So instead of using
`unimplemented!`, just unconditionally return an error here. These
errors are handled properly and cause Firecracker to exit gracefully
instead.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>