feat: update block perf tests to use fio json
Block tests were using fio logs to calculate bandwidth.
The issue is that the values in logs are in KB/sec and
we were just summing them up, so getting a sum of averages.
But what we want is just an average bandwidth for the test.
Now we use json output of fio which contains much more info
in a nice format. This info already includes all bandwidth
calculations. This change will also allow us to output
other metrics in the future if we want.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>