test: Remove `test_snapshot_compatibility`
This test serves no purpose, as all it does is take a snapshot without
verifying anything about it. In this sense, it is perfectly subsumed by
`test_balloon_snapshot`, which also creates snapshots with balloon
devices, but actually verifies they restore as well (something that is
not given in this test, because we do not wait for the guest to boot
before taking a snapshot).

A look at the commit history shows that this test used to be responsible
for verifying cross-release snapshot support for the balloon device, but
since we no longer support such a thing, it became superfluous.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>