tests: Call `wait_for_up` in `Microvm.restore_from_snapshot`
After restoring a snapshot, we are not waiting for the VM to finish
booting (it needs to have done that before we took a snapshot), but
having a `wait_for_up` here is still a good sanity check for ensuring
the VM is still responsive. So far our tests have been inconsistent on
whether to put a `wait_for_up` here. This makes them consistent by
always doing `wait_for_up` if a network device is available.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>