fix(aarch64): disable CONFIG_ARM64_ERRATUM_3194386 option
Disable CONFIG_ARM64_ERRATUM_3194386 option in the guest
kernels for aarch64. It is enabled by default
and makes `ssbs` feature to not be user space visible. This breaks
out integration tests.
This option is only indented for `Cortex-X4` and
`Neoverse-V3` cpus which we are not using, so disabling it
is safe.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>