unittest: net: Fix undefined behavior in `TestHelper::get_default`
This function used an unsound `mem::transmute` to invalidly extend a
lifetime with the goal of creating a self-referential struct. Sadly,
after returning from `get_default`, the `GuestMemoryMmap` gets moved to
a different stack frame, and so all references that this function
created become dangling pointers.

Fix this by not trying to create self-referential struct.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>