buildkite: eliminate `initial_steps`
The point of these is to run in parallel with the compilation steps, and
not block the wait step if they fail. However, since we use `depends_on:
null` for these, it doesn't matter whether they are initial or not, so
just add them via `add_step` with `decorate=False`, as we do for normal
steps.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>