buildkite: drop explicit `"depends_on": None` step entries
These were needed when the we still had a `wait` step (to overwrite the
`wait` step and instead start immediately), however since #4803 we do
not have a `wait` step anymore (instead everything that used to be after
the wait step now explicitly depends on the compilation steps). Thus,
having the `depends_on: null` entries for kani and style steps is no
longer needed (it also does not harm, but it looks ugly in buildkite's
canvas view).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>