test: style: fix black ignoring `integration_test/build` directory
When `black` is passed multiple directories, it will try to look for
`pyproject.toml` in the common ancestor of those paths. In our cases,
since we pass it `tests`, `.buildkite` and `tools`, that would be the
repository root. However, our pyproject.toml sits in `tests`, and thus
the `exclude` rule it defines is ignored (and it is this `exclude` rules
that overwrites the default behavior of excluding directories called
`build`).

Fix this by explicitly passing in a `--config` argument. Since `isort`
does not support this argument, pass it in a quite ugly via by adding it
to the pytest parameters.

Fixes: 4f55e491ee3bd08c21cf15db06162f4b4afccb04
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>