feat(virtio): add IovRingBuffer type
Add new `IovRingBuffer` ring buffer type that is tailored for holding
`struct iovec` objects that point to guest memory for IO. The
`struct iovec` objects represent the memory that the guest passed to
us as `Descriptors` in a VirtIO queue for performing some I/O operation.

We plan to use this type to describe the guest memory we have available
for doing network RX. This should facilitate us in optimizing the
reception of data from the TAP device using `readv`, thus avoiding a
memory copy.

Co-authored-by: Babis Chalios <bchalios@amazon.es>
Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>