refactor(net): replace rx_deferred_frame
Before `readv` change, `rx_deferred_frame` was used
to signal if there was present packet in the internal
rx buffer, so we would try to send it first before
attempting to read new one. With `readv` we don't
use internal buffer for RX packets anymore, so the
logic of `deferred` packet changed. Now any packet
is `deferred` when it is written to the guest memory.
After the write we try to notify guest about it, but
only if rate limiter allows it. To track this we
now use `deferred_rx_bytes` which holds the number of
bytes we yet to notify guest about. We can conveniently
use `Option<NonZeroUsize>` for this where `Some(...)`
will mean there are some `deferred` bytes we yet to
notify guest about and `None` will mean we are done and
can read the next packet.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>