fix(net): set tap offload features on restore
Tap offload features configuration was moved from the device creation
time to the device activation time by the following commit:

commit 1e5d3dba6a92358e3bee3ccb429ef383c479c588
Author: Nikita Zakirov <zakironi@amazon.com>
Date:   Fri Jan 19 15:48:21 2024 +0000

    fix(net): Apply only supported TAP offloading features

Since device activation code is only called on the boot path, the
features were not automatically configured on the restore path.
This change configures them on the restore path as well.

The change does not include a unit test as we do not have a mockable
interface for the tap device.
The change does not include an integration test as we have not yet found
a way to reproduce the issue using the existing test framework.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>