Revert readv optimization in virtio-net device
The changes introduced in IoVecBufferMut that allowed implementing readv
support for virtio-net regressed the performance of vsock device. This
is because now, when we create IoVecBufferMut objects we do a bunch (1
memfd_create and 3 mmap) of system calls. virtio-net device avoids the
performance issue because it creates a single IoVecBufferMut and re-uses
it, whereas vsock is creating a new IoVecBufferMut object for every
packet it receives.

We have a fix for this, which essentially creates a single
IoVecBufferMut that reuses for all the incoming vsock packets during a
connection. The problem with the fix is that it makes unit-tests really
unhappy and we need a significant amount of work to fix them.

So, revert the PR to have main in a clean state. We will fix the vsock
issues out-of-band and re-open the PR.

This reverts commits:
- bc0ba43d798c8322b33cc57251272ff5c4ae3560
- 667aba4536cecf86fabd1d6c1a6f0c0754447d6e
- 5d718d4d13bf21bd9bb317c97e87ca0a3451f05f
- 14e6e33ad329ad54232f1652375455b98df4adb8
- 1e4c632891178e00b9c614a0d5227c583a583110

from PR #4799

Signed-off-by: Babis Chalios <bchalios@amazon.es>