refactor(vsock): make vsock reuse packets
Add rx_packet and tx_packet members to the Vsock
type and reuse them for RX and TX processing. This removed
overhead of creating new packets for each request.

As an additional improvement, we split VsockPacket
type into 2 types: VsockPacketRx and VsockPacketTx
which allows us to separate logic for RX and TX
processing.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>