fix(ci): epoll() on pidfd to wait for Firecracker exit
Currently, we use psutil.pid_exists in a loop with a timeout of 10
seconds. This is racy and indeed some times we hit it in our CI.

Substitute this mechanism with calling epoll() on the pidfd of the
process instead. This should deterministically block until the process
exits. If there's something else wrong, we will hit the pytest timeout.

Signed-off-by: Babis Chalios <bchalios@amazon.es>