Combine SSHConnection._init_connection and Microvm.wait_for_ssh_up
We use `Microvm.wait_for_ssh_up` to wait for a booted guest's userland
to be initialized enough that it can process incoming SSH connections.
However, it turns out that we werent waiting the intended 10s, we were
waiting 3s. The reason for this is that before the `true` command in
`wait_for_ssh_up` is even sent to the guest, we go through the
`Microvm.ssh` property, which calls the `Microvm.ssh_face` function,
which constructs an `SSHConnection` object, whose construtor calls
`SSHConnection._init_connection`. And inside `_init_connection`, we
_already_ try to send a `true` command to the guest, to ascertain it is
up and running. However, here we do it in a retry loop (needed because
if port 22 is still closed, `ssh` exits with non-zero code and
"connection refuses") that does 20 retries at 0.15s intervals. Or tries
to 3 seconds. Most importantly, the retries in `_init_connection` are
done without timeouts, so even _if_ we actually hung during connection
establishment, `wait_for_ssh_up` wouldn't abort after 10 seconds,
because it would get stuck inside of `_init_connection`, which doesn't
know anything about a 10s timeout.

Fix all this up by merging `_init_connection` and `wait_for_ssh_up`.
Skip sending the `true` command in `wait_for_ssh_up`, and just call
`ssh_iface(0)` directly. Increment the retry time in `_init_connection`
to actually be 10s. And add a 10s timeout to the commands in
`_init_connection`.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>