test(net): relax timeouts
The main reason we have timeouts is so that we can get some useful
debugging information in case they do happen. This means the timesouts
only need to be strict enough to avoid running into the pytest timeout
instead. 10s/1s however still gets us spurious failures, since in the
non-performance tests contention on host networking resources (we use a
TON of networking namespaces) is pretty bad. So increase the timeouts to
something that still allows us to not run into the pytest timeout, while
hopefully returning spurious failures once and for all.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>