fix(fdt): Support FDT setup for CPU caches with high number of sets
Change the type of `CacheEntry`'s `number_of_sets` field from `u16` to
`u32`, to allow correctly parsing CPU cache information from sysfs and
successfully setting up the FDT on hosts with CPU caches with a number
of sets that is higher than `u16::MAX`.

This also makes a couple of `u16`->`u32` conversions redundant, which
we therefore remove.

Signed-off-by: Christos Katsakioris <ckatsak@cslab.ece.ntua.gr>
Signed-off-by: Filippos Tofalos <phtof@cslab.ece.ntua.gr>