fix(ci): condition for a non-existent process
`pidfd_open` will fail if there is not a process with the requested PID.
According to `man pidfd_open(2)`, it will return EINVAL when `PID` is
not valid and `ESRCH` when the `PID` does not exist. Right now, we were
checking only for the latter condition. Change the logic to also care
for the former, which materializes as an OSError exception with
errno == EINVAL.

Signed-off-by: Babis Chalios <bchalios@amazon.es>