fix: error handling in AML code generation
Right now, we include a few panics and assertions in the code that
generates AML bytecode for our ACPI tables.

Change the AML methods to explicitly return a Result type and substitute
these panics and assertions with error types. Also, change some type
methods that now need to return a Result as well.

Signed-off-by: Babis Chalios <bchalios@amazon.es>