fix: potential underflow in AML bytecode creation
acpi_tables::aml::AddressSpace<T> type is describing an address space
with a minimum address `min: T` and a maximum address `max: T`.
Currently, we do not perform any checks on these values when creating
the AddressSpace objects. This means that the starting address `min` can
be bigger than the last address `max`, which can cause underflows when
calculating the length of this address space, i.e. `max - min + 1`.

Add a check in the constructor methods of AddressSpace objects and
return an error when `min > max`.

Signed-off-by: Babis Chalios <bchalios@amazon.es>