test: ab: operate on directories instead of commit SHAs
The tools/ab_test.py script relies on precompiled binaries existing in
well-defined locations, and refuses to compile firecracker itself if
they're missing. It derives these locations from commit SHAs, so
conceptually, it makes more sense to cut out this middle step and just
directly pass in directories instead of SHAs (with the expectation that
the directories contain firecracker and jailer binaries).

However, the commit SHAs were still used to print the commit ranges over
which A/B-tests were done. It turns out that this was not working in the
vast majority of cases though, as the commit log printing logic did not
contain all the resolution logic that goes into compilation step (e.g.
for parsing revisions and such). So just remove that part.

In the EMF metrics, we now tag each report with the directory path
instead of the commit SHA. For the post-merge runs, this makes no
difference, as the SHA is part of the path.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>