devtool: Add flag to build to allow compiling arbitrary revisions
Passing `--rev` to `tools/devtool build` will do a checkout of the
specified revision into a temporary directory, compile firecracker into
it, and then copy over the final binaries into build/$revision.

This has advantages over our current approach for compiling A/B
revisions, as by only copying the final binaries into build/$revision we
save a lot of bandwidth when uploading artifacts for transfer between
different buildkite steps (as we no longer include gigabytes of
compilation artifacts).

This now means that all revisions are compiled in the same docker
container, however due to the rust-toolchain.toml this should not have
an impact on cross-toolchain testability, and has the advantage that the
cargo cache is shared (meaning we're hitting crates.io for downloads
less).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>