test: stop compiling firecracker inside A/B-tests
In the pre-PR A/B-tests we were compiling Firecracker. Instead,
explicitly rely on the precompiled binaries that we set up in the shared
buildkite build step. This has the slight downside that it makes it
harder to run these tests locally, as now you need to explicitly
pre-compile the binaries, but arguably running these vulnerability
A/B-tests locally doesnt make sense anyway, because they specifically
test the security configuration on our supported .metals.

While we're at it, rename the ab_* utility functions to make it obvious
that they expect pre-compiled binaries.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>