test: ab: Add function for A/B-Tests across precompiled binaries
Currently, we have two types of pre-PR A/B-Tests: Those that depend on
the repository as a whole (e.g. `cargo audit`, which checks properties
of Cargo.toml), and those that depend on the pre-compiled firecracker
binaries.

At the moment, these tests can use the same infrastructure, because our
buildkite shared build step will pre-compile binaries and then upload
the entire repository as an artifact, and so both types of A/B-tests
will find what they're looking for in this artifact and end up being
happy. However, this model makes it very difficult to manually run
A/B-tests locally (one will need to setup clones of the git repository
at specific locations and then pre-compile firecracker with specific
flags inside them). This patch series has the goal to simplify these
manual A/B-tests, but the cost is that the pre-PR A/B tests will require
different infrastructure.

Thus, prepare by providing different functions for each of these types
of A/B-test to use.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>