chore(test): Double refill time for RX rate limiter
Although we increased the refill time in the past in commit
347f877ec2a9 ("unittest: net: eliminate ReadTapMock::MockFrame") to
avoid intermittent issues, it still fails rarely (maybe due to resource
contention). We are not sure the enough refill time, but it's worth a
try and may reduce the chance of kicking the retry of the whole unit
tests manually.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>